requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
pydantic==2.5.2

# AI/ML Dependencies
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
pillow==10.1.0

# Medical Image Processing
monai==1.3.0
nibabel==5.1.0
SimpleITK==2.3.1

# Database
psycopg2-binary==2.9.9
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1 